requirements.txt
requests
Pillow
pyTelegramBotAPI
python-dotenv
pytz